wk_dev

+++ Walkaroo/LatestProd(12-08-2026)/WrapperService/src/main/java/com/tekclover/wms/core/model/spark/PalletizationDashBoardResponse.java
@@ -3,6 +3,7 @@ package com.tekclover.wms.core.model.spark;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Data
@@ -19,5 +20,5 @@ public class PalletizationDashBoardResponse {
     private Timestamp startTime;
     private Timestamp endTime;
     private Long leadTimeMinutes;
-    private Long productivityPerHour;
+    private BigDecimal productivityPerHour;
 }
