+++ Walkaroo/LatestProd(12-08-2026)/WrapperService/src/main/java/com/tekclover/wms/core/model/spark/PalletizationDashBoardResponseList.java
@@ -2,6 +2,7 @@ package com.tekclover.wms.core.model.spark;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Data
@@ -18,6 +19,6 @@ public class PalletizationDashBoardResponseList {
     private Timestamp startTime;
     private Timestamp endTime;
     private Long leadTimeMinutes;
-    private Long productivityPerHour;
+    private BigDecimal productivityPerHour;
 
 }
